pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "riglogix-router"
version = "0.1.0"
description = "RLR-4SS — Drilling opportunity radar and commercial copilot"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "apscheduler>=3.10.4",
    "jinja2>=3.1.4",
]

[tool.setuptools.packages.find]
where = ["."]
